Move current Icon component into _DeprecatedIcon
