+++ src/components/Datepicker/components/DatepickerNavigation.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Icon from '../../Icon';
+import Icon from '../../_DeprecatedIcon';
 import IconButton from '../../IconButton';
 import { StyledDatepickerNavigation } from '../styles';
