+++ src/components/Uploader/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import Icon from '../Icon';
+import Icon from '../_DeprecatedIcon';
 import IconButton from '../IconButton';
 import Spinner from '../Spinner';
 import UploaderImageItem from '../UploaderImageItem';
