+++ src/components/Uploader/styles.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import FormElementWrapper from '../FormComponents/FormElementWrapper';
-import Icon from '../Icon';
+import Icon from '../_DeprecatedIcon';
 import UiText from '../UiText';
 
 import { theme } from '../../theme';
