+++ src/components/_DeprecatedIcon/index.js
